Abstract loading a JSON/TOML/KDL file
