+++ packages/sync-server/app/utils/loadSettingFile.js
@@ -1,0 +1,22 @@
+import { promises as fs } from 'fs';
+import TOML from '@iarna/toml';
+
+// it does work, but eslint doesn't like it
+// eslint-disable-next-line import/no-unresolved
+import { parse as parseJiK } from '@bgotink/kdl/json';
+
+export async function loadSettingFile(filepath) {
+  const file = (await fs.readFile(filepath)).toString();
+  let value;
+  if (filepath.endsWith('.json')) {
+    value = JSON.parse(file);
+  } else if (filepath.endsWith('.toml')) {
+    value = TOML.parse(file);
+  } else if (filepath.endsWith('.kdl')) {
+    value = parseJiK(file);
+  } else {
+    throw new Error('File format not supported');
+  }
+
+  return value;
+}
